requirements.txt
beautifulsoup4==4.10.0
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
cssselect==1.1.0
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.3.0
google==3.0.0
gTTS==2.2.3
idna==3.2
jieba3k==0.35.1
joblib==1.0.1
lxml==4.6.3
newspaper3k==0.2.8
nltk==3.6.3
numpy==1.21.2
Pillow==8.3.2
PyAudio==0.2.11
python-dateutil==2.8.2
PyYAML==5.4.1
regex==2021.9.30
requests==2.26.0
requests-file==1.5.1
scikit-learn==1.0
scipy==1.7.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
threadpoolctl==3.0.0
tinysegmenter==0.3
tldextract==3.1.2
tqdm==4.62.3
urllib3==1.26.7
